made both FrameSet and FrameSetCollection classes private
